requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
python-telegram-bot[job-queue]==20.8
python-dotenv==1.0.1
dj-database-url==2.1.0
